Updated User and Cart Controller

+++ backend/controller/cartController.js
@@ -21,17 +21,15 @@ const addCart = async(req, res) => {
 
     const refreshToken = cookies.jwt;
 
-    const foundUser = await User.findOne({ refreshToken }).exec();
+    const foundUser = await User.findOne({refreshToken: refreshToken }).exec();
 
     if (!foundUser) {
-        //res.clearCookie('jwt', { httpOnly: true, sameSite: 'None', secure: true });
-        res.clearCookie('jwt', { httpOnly: true, secure: true, sameSite: 'None'});
+        res.clearCookie('jwt', { httpOnly: true, sameSite: 'None', secure: true });
+        //res.clearCookie('jwt', { httpOnly: true, secure: true, sameSite: 'None'});
         return res.sendStatus(204);
     }
 
     const owner = foundUser._id;
-
-    console.log(owner)
 
     const itemName = req.body.name;
 
@@ -39,7 +37,7 @@ const addCart = async(req, res) => {
 
     try {
         const cart = await CartDAO.getCart(owner);
-        const item = await Item.findOne({ Name: itemName });
+        const item = await Item.findOne({ name: itemName });
         if (!item) {
             res.status(404).send({ message: "item not found" });
             return;
@@ -77,7 +75,7 @@ const addCart = async(req, res) => {
         {
             //no cart exists, create one
             const newCartData= {
-                _id:owner,
+                owner:owner,
                 items: [{ itemId, name, quantity, price }],
                 bill: quantity * price,
             }
@@ -105,7 +103,7 @@ const deleteCartItem = async (req, res) => {
 
     const refreshToken = cookies.jwt;
 
-    const foundUser = await User.findOne({ refreshToken }).exec();
+    const foundUser = await User.findOne({refreshToken: refreshToken }).exec();
 
     if (!foundUser) {
         //res.clearCookie('jwt', { httpOnly: true, sameSite: 'None', secure: true });
@@ -154,7 +152,7 @@ const deleteCart = async (req, res) => {
 
     const refreshToken = cookies.jwt;
 
-    const foundUser = await User.findOne({ refreshToken }).exec();
+    const foundUser = await User.findOne({refreshToken: refreshToken }).exec();
 
     if (!foundUser) {
         //res.clearCookie('jwt', { httpOnly: true, sameSite: 'None', secure: true });
@@ -166,7 +164,7 @@ const deleteCart = async (req, res) => {
     
     try 
     {
-        const result = await Cart.deleteOne({ owner });
+        const result = await Cart.deleteOne({ owner: owner });
         if (result.deletedCount == 1) {
             res.sendStatus(200);
         }
